backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
sqlalchemy
pydantic[email]
pydantic-settings
python-dotenv
boto3
bcrypt==3.2.0
